resolved suggestions from last push
